pyproject.toml
[tool.poetry]
name = "PCVwP"
version = "0.1.0"
description = ""
authors = ["Kotaro SONODA <kotaro1976@gmail.com>"]

[tool.poetry.dependencies]
python = ">=3.10,<3.11"
Pillow = "^9.1.0"
ipykernel = "^6.13.0"
numpy = "^1.22.3"
matplotlib = "^3.5.1"
SciencePlots = "^1.0.9"
ipympl = "^0.9.0"
scipy = "^1.8.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
